Give the api back its blocking serve and split MailDevice out
The api was the last service and its Serve blocked, so main died with it and
docker restarted the container. That was traded for a background goroutine
with log.Fatalf and a bare select{} in main only because the inbound server
used to block ahead of it waiting for a certificate. Nothing blocks there any
more, so the original shape is back.

MailDevice moves to its own module and the comments the e2e tests already
cover are dropped.
