+++ backend/rest/api.go
@@ -147,12 +147,7 @@ func (a *Api) Start() error {
 	}
 	l := netutil.LimitListener(listener, 1000)
 	log.Printf("Started backend (%s)\n", a.socket)
-	go func() {
-		if err := srv.Serve(l); err != nil {
-			log.Fatalf("backend stopped: %v", err)
-		}
-	}()
-	return nil
+	return srv.Serve(l)
 }
 
 func (a *Api) Status(_ http.ResponseWriter, req *http.Request) (interface{}, error) {
